+++ config.php
@@ -2,6 +2,7 @@
 
 use humhub\modules\darkMode\Events;
 use humhub\widgets\LayoutAddons;
+use humhub\modules\admin\models\forms\DesignSettingsForm;
 use humhub\widgets\NotificationArea;
 use humhub\components\ModuleManager;
 use yii\base\Event;
@@ -13,6 +14,7 @@ return [
     'events' => [
         ['class' => NotificationArea::class, 'event' => NotificationArea::EVENT_INIT, 'callback' => [Events::class, 'onNotificationAddonInit']],
         ['class' => LayoutAddons::class, 'event' => LayoutAddons::EVENT_INIT, 'callback' => [Events::class, 'onLayoutAddonsInit']],
+        ['class' => DesignSettingsForm::class, 'event' => DesignSettingsForm::EVENT_AFTER_VALIDATE, 'callback' => [Events::class, 'onDesignSettingForm']],
         ['class' => ModuleManager::class, 'event' => ModuleManager::EVENT_AFTER_MODULE_ENABLE, 'callback' => [Events::class, 'onAfterModuleEnabled']],
         ['class' => ModuleManager::class, 'event' => ModuleManager::EVENT_BEFORE_MODULE_DISABLE, 'callback' => [Events::class, 'onBeforeModuleDisabled']]
     ]