+++ messages/it/admin.php
@@ -1,10 +1,10 @@
 <?php
 
 return [
-    '<strong>Dark Mode</strong> module configuration' => '',
-    'Adds a Dark Mode to HumHub' => '',
-    'Current base theme: ' => '',
-    'Dark Mode' => '',
-    'Dark Theme' => '',
-    'The stylesheet of the selected theme will be used for the dark mode.' => '',
+    '<strong>Dark Mode</strong> module configuration' => 'Configurazione del modulo <strong>Modalità scura</strong>',
+    'Adds a Dark Mode to HumHub' => 'Aggiunge la modalità scura a HumHub',
+    'Current base theme: ' => 'Tema di base corrente ',
+    'Dark Mode' => 'Modalità scura',
+    'Dark Theme' => 'Tema scuro',
+    'The stylesheet of the selected theme will be used for the dark mode.' => 'Il foglio di stile del tema selezionato verrà utilizzato per la modalità scura.',
 ];